Adjust explanation when user loses EOA
When a user loses access to an EOA, it was suggested to add multiple owners to the Safe. However, this can be confusing to the users as the Gnosis Pay Safe cannot have multiple owners.

Instead, we should recommend a setup where Safe with multiple owners is the owner of the Gnosis Pay Safe. This is currently not supported by our WebApp, but it is supported by many partners.

Therefore, this is just a simple adjustment to the given recommendation (as a more in-depth one would require mentioning a non-EOA setup).